Extract org report writers to org/writers/ subpackage
Move 16 write_org_report_* functions and supporting helpers
(~1,850 LOC) from generator.py to org/writers/__init__.py.
Re-imported in generator.py for backwards compatibility.

This is the critical extraction for the v3.4.0 trending feature —
new trending output code will land directly in org/writers/.
